Fix "zoomto" issue on added layers.
ESRI layers added from search did not have their BBOX set properly.

This ensures that the bbox is set properly when the layer is loaded
at start up (Loading a map, create map from layer) and when the layer
is added to the map from the search (Add Layers Dialog).

refs: BEX-695